WTA #82: updated dto and mapper
