+++ core/common/src/main/java/com/jacob/wakatimeapp/core/common/data/local/WakaTimeAppDB.kt
@@ -3,8 +3,11 @@ package com.jacob.wakatimeapp.core.common.data.local
 import arrow.core.Either
 import com.jacob.wakatimeapp.core.common.data.dtos.ExtractedDataDTO
 import com.jacob.wakatimeapp.core.common.data.local.dao.ApplicationDao
+import com.jacob.wakatimeapp.core.common.data.local.entities.DayEntity
+import com.jacob.wakatimeapp.core.common.data.local.entities.ProjectPerDay
 import com.jacob.wakatimeapp.core.models.DetailedDailyStats
 import com.jacob.wakatimeapp.core.models.Error
+import com.jacob.wakatimeapp.core.models.Range
 import javax.inject.Inject
 import javax.inject.Singleton
 import kotlinx.datetime.LocalDate
@@ -13,11 +16,17 @@ import kotlinx.datetime.LocalDate
 class WakaTimeAppDB @Inject constructor(
     private val applicationDao: ApplicationDao,
 ) {
-    suspend fun getStatsForDay(date: LocalDate) = applicationDao.getStatsForDay(date)
+    suspend fun getStatsForDay(date: LocalDate): Either<Error, List<DayEntity>> = Either.catch {
+        applicationDao.getStatsForDay(date)
+    }.mapLeft { Error.DatabaseError.UnknownError("could not get stats for day ($date)", it) }
 
-    suspend fun getStatsForProject(name: String) = applicationDao.getStatsForProject(name)
+    suspend fun getStatsForProject(name: String): Either<Error, List<ProjectPerDay>> = Either.catch {
+        applicationDao.getStatsForProject(name)
+    }.mapLeft { Error.DatabaseError.UnknownError("could not get stats for project ($name)", it) }
 
-    suspend fun getDateRangeInDb() = applicationDao.getDateRangeInDb()
+    suspend fun getDateRangeInDb(): Either<Error, Range> = Either.catch {
+        applicationDao.getDateRangeInDb()
+    }.mapLeft { Error.DatabaseError.UnknownError("could not get range from db", it) }
 
     suspend fun insertExtractedData(extractedDataDTO: ExtractedDataDTO): Either<Error, Unit> = Either.catch {
         applicationDao.insertStatesForDay(extractedDataDTO)
