+++ details/src/main/java/com/jacob/wakatimeapp/details/data/ProjectDetailsPageAPI.kt
@@ -2,7 +2,6 @@ package com.jacob.wakatimeapp.details.data
 
 import com.jacob.wakatimeapp.core.common.data.dtos.ProjectDetailsDTO
 import com.jacob.wakatimeapp.details.data.dtos.TotalTimeForProjectDTO
-import com.jacob.wakatimeapp.login.data.dtos.DetailedProjectStatsDTO
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Header
@@ -15,13 +14,13 @@ internal interface ProjectDetailsPageAPI {
         @Query("q") projectName: String,
     ): Response<ProjectDetailsDTO>
 
-    @GET("/api/v1/users/current/summaries")
-    suspend fun getStatsForProject(
-        @Header("Authorization") token: String,
-        @Query("project") projectName: String,
-        @Query("start") start: String,
-        @Query("end") end: String,
-    ): Response<com.jacob.wakatimeapp.login.data.dtos.DetailedProjectStatsDTO>
+//    @GET("/api/v1/users/current/summaries")
+//    suspend fun getStatsForProject(
+//        @Header("Authorization") token: String,
+//        @Query("project") projectName: String,
+//        @Query("start") start: String,
+//        @Query("end") end: String,
+//    ): Response<DetailedProjectStatsDTO>
 
     @GET("/api/v1/users/current/all_time_since_today")
     suspend fun getTotalTimeForProject(
