+++ login/src/main/java/com/jacob/wakatimeapp/login/domain/usecases/LoadExtractedDataIntoDbUC.kt
@@ -5,11 +5,9 @@ import arrow.core.left
 import com.jacob.wakatimeapp.core.common.data.dtos.ExtractedDataDTO
 import com.jacob.wakatimeapp.core.common.data.local.WakaTimeAppDB
 import com.jacob.wakatimeapp.core.models.Error
-import com.jacob.wakatimeapp.core.models.Error.DatabaseError.UnknownError
 import javax.inject.Inject
 import javax.inject.Singleton
 import kotlinx.serialization.json.Json
-import timber.log.Timber
 
 @Singleton
 internal class LoadExtractedDataIntoDbUC @Inject constructor(
@@ -23,11 +21,6 @@ internal class LoadExtractedDataIntoDbUC @Inject constructor(
 
         val extractedDataDTO = json.decodeFromString<ExtractedDataDTO>(String(bytes))
 
-        return Either.catch {
-            wakaTimeAppDB.insertExtractedData(extractedDataDTO)
-        }.mapLeft {
-            Timber.e("could not insert extracted data into db", it)
-            UnknownError("could not insert extracted data into db", it)
-        }
+        return wakaTimeAppDB.insertExtractedData(extractedDataDTO)
     }
 }
